feat(eslint-config): add prefer-arrow-functions rule
This commit introduces the 'prefer-arrow-functions' rule from the
eslint-plugin-prefer-arrow-functions package. This rule enforces the use
of arrow functions where possible, improving code consistency and
readability. The rule is set to 'warn' level and is applied to all files.

BREAKING CHANGE: This may cause existing code that does not use arrow
functions where possible to throw warnings.